cpu interrupts: handle VBLANK interrupt and don't forget to reenable ime flag after handling any interrupt on the next instruction cycle (during cpu state CPU_EXECUTION_STATE_EXECUTE)

+++ gameboy/cpu_registers.go
@@ -129,6 +129,6 @@ func (c *CPU) GetIEFlag() uint8 {
 	return c.bus.Read(IE_REGISTER)
 }
 
-func (c *CPU) setIEFlag(value uint16) {
+func (c *CPU) setIEFlag(value uint8) {
 	c.bus.Write(IE_REGISTER, byte(value))
 }
